Move `FileBase` from `core` to `fs` (so it can be imported to wing).
