+++ libs/wingsdk/src/fs/index.ts
@@ -1,2 +1,3 @@
+export * from "./filebase";
 export * from "./json-file";
 export * from "./text-file";
